refactor(cli): per-verb file split + flip pnpm test to vitest (T5.2)
Closes Phase 5. Splits the 13 session verbs into individual files
under packages/cli/src/commands/ and flips the workspace test runner
to Vitest exclusively.

Verb split:
- 13 verbs now each live in packages/cli/src/commands/{start,list,
  show,checkpoint,pause,end,attach,continue,archive,restore,delete,
  search,events}.ts (every file < 100 LOC, most under 25).
- _shared.ts hosts the Command function-shape, requireSession()
  narrower, and the runLifecycle() helper that checkpoint/pause/end
  share. index.ts exports a verb-to-handler map.
- runtime.ts collapses from 428 → 131 LOC and now just dispatches
  via sessionVerbs[verb]. rebuild + seed (top-level commands) stay
  inline. usage / sessionsUsage strings are unchanged.

Test runner flip:
- test/healthcheck.test.js + test/integrations.test.js ported to
  vitest under test/*.test.ts (3 + 17 tests, byte-identical
  coverage).
- New root vitest.config.ts picks them up via include:
  ["test/**/*.test.ts"], passWithNoTests: true.
- package.json `test` simplifies to:
    pnpm -r run build && pnpm -r run test:vitest && vitest run
  — drops test:node, test:vitest, test:sessions aliases.
- scripts/check-test-count.mjs tolerates zero node:test files and
  now counts vitest in two phases: workspace (`pnpm -r exec vitest`)
  plus root (`pnpm exec vitest run`).

Snapshot tests (T5.2 acceptance):
- packages/cli/tests/snapshots.test.ts pins help screens (top-level
  + sessions) and the --json output shapes for sessions.list /
  sessions.events. Unknown-verb dispatch also pinned.

Coverage: 249 tests (node:test=0, vitest=249), floor 177.
