+++ package.json
@@ -12,10 +12,7 @@
     "dashboard": "pnpm --filter @librarian/mcp-server serve",
     "rebuild": "pnpm --filter @librarian/cli rebuild",
     "seed": "pnpm --filter @librarian/cli seed",
-    "test": "pnpm -r run build && pnpm run test:node && pnpm run test:vitest",
-    "test:node": "node --no-warnings --test packages/*/tests/*.test.js test/*.test.js",
-    "test:sessions": "node --no-warnings --test packages/core/tests/sessions.test.js packages/mcp-server/tests/sessions.mcp.test.js packages/mcp-server/tests/sessions.http.test.js test/integrations.test.js",
-    "test:vitest": "pnpm -r run test:vitest",
+    "test": "pnpm -r run build && pnpm -r run test:vitest && vitest run",
     "smoke": "node --no-warnings scripts/smoke-test.js",
     "healthcheck": "node --no-warnings scripts/healthcheck.js",
     "lint": "eslint .",
