+++ packages/cli/src/commands/end.ts
@@ -1,0 +1,4 @@
+import { type Command, runLifecycle } from "./_shared.js";
+
+export const end: Command = (store, positionals, flags) =>
+  runLifecycle(store, "end", positionals[0], flags);
