+++ packages/cli/src/commands/pause.ts
@@ -1,0 +1,4 @@
+import { type Command, runLifecycle } from "./_shared.js";
+
+export const pause: Command = (store, positionals, flags) =>
+  runLifecycle(store, "pause", positionals[0], flags);
